feat(backend): Replace zod.number with integer

+++ packages/backend/src/Validations/Zod/ZodPagination.ts
@@ -1,8 +1,8 @@
 import { z } from 'zod/v4'
 
 const ZodPagination = z.object({
-	page: z.coerce.number().min(0).optional(),
-	limit: z.coerce.number().min(0).optional(),
+	page: z.coerce.number().int().min(0).optional(),
+	limit: z.coerce.number().int().min(0).optional(),
 })
 
 export default ZodPagination
